+++ src/main/java/com/qcloud/Module/APIGateway.java
@@ -1,0 +1,12 @@
+package com.qcloud.Module;
+
+/*
+ * see document: <a
+ * href="https://cloud.tencent.com/document/product/628">
+ * https://cloud.tencent.com/document/product/628</a>
+ */
+public class APIGateway extends Base {
+    public APIGateway() {
+        serverHost = "apigateway.api.qcloud.com";
+    }
+}